services/tiles/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Redis caching
redis==5.0.1
hiredis==2.3.2

# HTTP client
httpx==0.26.0

# Image processing (for AI depth maps)
Pillow==10.2.0
numpy==1.26.3

# Utilities
python-dotenv==1.0.0
python-json-logger==2.0.7
